requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
PyMySQL==0.9.3
requests==2.22.0